requirements.txt
atomicwrites==1.3.0
attrs==19.2.0
boto3==1.10.19
botocore==1.13.19
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-cors-headers==3.2.0
django-rest-auth==0.9.5
django-storages==1.7.2
djangorestframework==3.10.3
docutils==0.15.2
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
pep8==1.7.1
Pillow==6.2.1
pkg-resources==0.0.0
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.2.0
python-dateutil==2.8.0
python-magic==0.4.15
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
six==1.12.0
South==1.0.2
sqlparse==0.3.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
